CFY-6288 Use "error" as the error message key (#67)
* CFY-6288 Use "error" as the error message key

Both the plugin and the service need to use the same key for error
messages, to display errors correctly. Service was using "message",
while plugin was using "error". Let's use "error" in both.

* CFY-6288 Test the error response format

Existing test that test error responses, now also assert on the
response format

* Make the "malformed data" error response have a message

The "this is not json" response can contain a message, not just be
empty - we have a dedicated exception class for that